+++ README.md
@@ -1,2 +1,4 @@
-# otus_spring_homework5
-homework 5 for Otus springFramework course
+# otus spring homework 7
+homework 7 for Otus springFramework course
+
+spring-data-jpa
