--- src/main/java/ru/ezikvice/springotus/homework5/dao/AuthorRepository.java
@@ -1,16 +1,0 @@
-package ru.ezikvice.springotus.homework5.dao;
-
-import ru.ezikvice.springotus.homework5.domain.Author;
-
-import java.util.Set;
-
-public interface AuthorRepository {
-
-    int count();
-
-    void insert(Author author);
-
-    Author findById(int id);
-
-    Set<Author> findByName(String name);
-}
