--- src/main/java/ru/ezikvice/springotus/homework5/dao/GenreRepository.java
@@ -1,11 +1,0 @@
-package ru.ezikvice.springotus.homework5.dao;
-
-import ru.ezikvice.springotus.homework5.domain.Genre;
-
-public interface GenreRepository {
-    void set(Genre genre);
-
-    Genre findById(int id);
-
-    int count();
-}
