--- src/main/java/ru/ezikvice/springotus/homework5/service/GenreService.java
@@ -1,12 +1,0 @@
-package ru.ezikvice.springotus.homework5.service;
-
-import ru.ezikvice.springotus.homework5.domain.Genre;
-
-public interface GenreService {
-
-    int count();
-
-    void add(Genre genre);
-
-    Genre find(int genreId);
-}
