+++ src/test/java/ru/ezikvice/springotus/homework7/AuthorDaoTests.java
@@ -1,4 +1,4 @@
-package ru.ezikvice.springotus.homework5;
+package ru.ezikvice.springotus.homework7;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -9,26 +9,25 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.test.context.junit4.SpringRunner;
-import ru.ezikvice.springotus.homework5.dao.AuthorRepositoryJdbc;
-import ru.ezikvice.springotus.homework5.domain.Author;
+import ru.ezikvice.springotus.homework7.dao.AuthorRepository;
+import ru.ezikvice.springotus.homework7.domain.Author;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @DataJpaTest
-@ComponentScan("ru.ezikvice.springotus.homework5.dao")
+@ComponentScan("ru.ezikvice.springotus.homework7.dao")
 public class AuthorDaoTests {
 
     @Autowired
     TestEntityManager em;
 
     @Autowired
-    private AuthorRepositoryJdbc rep;
+    private AuthorRepository rep;
 
     @Test
     public void findingAuthorByName() {
         Author testNameAuthor = new Author("Test Name Author");
         em.persist(testNameAuthor);
-//        Set<Author> foundAuthors = service.findByName("Test Name Author");
         Assert.assertTrue(rep.findByName("Test Name Author").contains(testNameAuthor));
     }
 }
