+++ src/test/java/ru/ezikvice/springotus/homework7/AuthorTests.java
@@ -1,4 +1,4 @@
-package ru.ezikvice.springotus.homework5;
+package ru.ezikvice.springotus.homework7;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -6,8 +6,8 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
-import ru.ezikvice.springotus.homework5.domain.Author;
-import ru.ezikvice.springotus.homework5.service.AuthorService;
+import ru.ezikvice.springotus.homework7.domain.Author;
+import ru.ezikvice.springotus.homework7.service.AuthorService;
 
 import java.util.Set;
 
@@ -29,7 +29,7 @@ public class AuthorTests {
 
     @Test
     public void findingAuthorByName() {
-        Author testNameAuthor = new Author( "Test Name Author");
+        Author testNameAuthor = new Author("Test Name Author");
         service.add(testNameAuthor);
         Set<Author> foundAuthors = service.findByName("Test Name Author");
         Assert.assertTrue(foundAuthors.contains(testNameAuthor));
