+++ src/test/java/ru/ezikvice/springotus/homework7/GenreDaoTests.java
@@ -1,4 +1,4 @@
-package ru.ezikvice.springotus.homework5;
+package ru.ezikvice.springotus.homework7;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -9,25 +9,24 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.test.context.junit4.SpringRunner;
-import ru.ezikvice.springotus.homework5.dao.GenreRepositoryJdbc;
-import ru.ezikvice.springotus.homework5.domain.Author;
-import ru.ezikvice.springotus.homework5.domain.Genre;
+import ru.ezikvice.springotus.homework7.dao.GenreRepository;
+import ru.ezikvice.springotus.homework7.domain.Genre;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @DataJpaTest
-@ComponentScan("ru.ezikvice.springotus.homework5.dao")
+@ComponentScan("ru.ezikvice.springotus.homework7.dao")
 public class GenreDaoTests {
 
     @Autowired
     TestEntityManager em;
 
     @Autowired
-    private GenreRepositoryJdbc rep;
+    private GenreRepository rep;
 
     @Test
     public void findingGenreById() {
-         Genre testGenre = new Genre("Testing Software", "It`s all about us");
+        Genre testGenre = new Genre("Testing Software", "It`s all about us");
         em.persist(testGenre);
         Assert.assertEquals(rep.findById(testGenre.getId()), testGenre);
     }
